Moved the message queue to the MQTT Agent
