+++ exe/cpp_agent_template/Agent.hpp
@@ -23,7 +23,6 @@ class Agent : public Utils {
     string app_name;
 
     void process_message(const json::object &message);
-    void enqueue_message(const json::object &message);
 
     json::object parse_json(const string text);
 
@@ -40,5 +39,6 @@ class Agent : public Utils {
 
     public:
 
+    virtual void process_next_message() {}
     virtual void publish(json::object &message) {}
 };
