+++ exe/cpp_agent_template/MqttAgent.hpp
@@ -3,6 +3,10 @@
 #include <iostream>
 #include <boost/json.hpp>
 #include <mqtt/async_client.h>
+#include <thread>
+#include <queue>
+#include <future>
+
 #include "Agent.hpp"
 #include "ReferenceMessageHandler.hpp"
 
@@ -16,6 +20,16 @@ class MqttAgent : public Agent {
     std::shared_ptr<mqtt::async_client> mqtt_client;
 
     bool running = true;
+    bool busy = false;
+
+
+    // Message queue gets checked once per second
+    std::future<void> queue_future;
+    void check_queue();
+
+    // A FIFO queue of messages from the bus
+    queue<json::object> message_queue;
+
 
   public:
 
@@ -27,4 +41,6 @@ class MqttAgent : public Agent {
 
     /* send output to the message bus */
     void publish(json::object &message) override;
+
+    void process_next_message() override;
 };
